Make @@warmup view available on both Plone site and Zope app root:
On the Zope app root, it will find and warm up all Plone sites.
This allows to simplify constructing the URL to the warmup view
(e.g. during buildout) because if the view is invoked on the
Zope app root, the ID of the Plone site doesn't need to be known.