Don't attempt to launch a monitor server in bin/instance debug:
We only want to launch a monitor server (and a corresponding warmup
request) in cases where the instance that's being started actually
started an HTTP server.

That is not the case for bin/instance debug, bin/instance run or
bin/instance <zopectl> invocations, and we therefore need to
recognize these.